CelluloseBuilder

Documentation for CelluloseBuilder.

CelluloseBuilder.PBC

CelluloseBuilder.PBC

CelluloseBuilder.PBC

CelluloseBuilder.PBC

CelluloseBuilder.cellulosebuilder

CelluloseBuilder.cleaning_tmpfiles

CelluloseBuilder.fractional2cartesian

CelluloseBuilder.fragPDBs

CelluloseBuilder.isinverted

CelluloseBuilder.lattice2basis

CelluloseBuilder.lattice2basis

CelluloseBuilder.monolayer

CelluloseBuilder.pbcXYZ

CelluloseBuilder.pbcdata

CelluloseBuilder.pbcdata

CelluloseBuilder.translate

CelluloseBuilder.translate

CelluloseBuilder.translate

CelluloseBuilder.writePDB

CelluloseBuilder.writeXYZ

CelluloseBuilder.z_expansion

CelluloseBuilder.PBC — Method

PBC(xsize::Int64, ysize::Int64, zsize::Int64; phase="Iβ", pbc=nothing)

Returns (ncells, lattice) where ncells is the number of unit cells along x, y and z axes and lattice is the unit cell vector of this cell. This function is used to define the periodic boundary conditions required to build the cellulose crystal.

CelluloseBuilder.PBC — Method

PBC(monomers::Int64, chains::Int64; phase="Iβ", layer="center")

Define the crystallographic space to craft a cellulose layer with a degree of β-glucose units (monomers) and a number of cellulose chains (chains).

CelluloseBuilder.PBC — Method

PBC(monomers::Int64; phase="Iβ")

Define the crystallographic space to build the cellulose fibril crystal with a degree of β-glucose units (monomers).

CelluloseBuilder.PBC — Method

PBC(xyzsizes::Vector{Int64}; phase="Iβ", pbc=nothing)

See PBC(xsize::Int64, ysize::Int64, zsize::Int64; phase="Iβ", pbc=nothing).

CelluloseBuilder.cellulosebuilder — Method

cellulosebuilder(monomers::Int64; phase="Iβ", fibril=nothing, ...)

Builds a cellulose fibril with a given number of cellobiose units based on the number of monomers.

CelluloseBuilder.cleaning_tmpfiles — Method

cleaning_tmpfiles()

Move the exported files to the work directory and clean all the *.pdb, *.tcl, and *xyz on temporary files inside the temp folder (it can be different on Linux, Windows and MacOS).

CelluloseBuilder.fractional2cartesian — Method

fractional2cartesian(unitcell::Vector{Int64}, phase::String)

Convert the fractional unit cell coordinates to the cartesian coordinates. The return is the x, y, and z cartesian coordinates.

CelluloseBuilder.fragPDBs — Method

fragPDBs(xyzfile::String, selection::String; vmd="vmd")

Convert each fragment on XYZ file on an unique raw PDB file. The filename will be exported, so there is a way to edit the PDB data.

CelluloseBuilder.isinverted — Method

isinverted(pdbname::String; vmd="vmd")

Check if the sugar monomer is inverted or not on two atoms. The default return checks if the first monomer of the cellulose chain is upside down.

CelluloseBuilder.lattice2basis — Method

lattice2basis(lattice::Vector{Int64}, phase::String)

CelluloseBuilder.lattice2basis — Method

lattice2basis(lattice::Vector{Int64}, parameters::Vector{Vector{Float64}})

Convert the lattice vector to the basis vectors of the unit cell on the cartesian space.

CelluloseBuilder.monolayer — Method

monolayer(xyzsizes::Vector{Int64}; phase="Iα", structure="monolayer")

Returns the fragments selection needed to the monolayer structure.

CelluloseBuilder.pbcXYZ — Method

pbcXYZ(nfrag::Int64, xsize::Int64, ysize::Int64; phase="Iβ", pbc=nothing, xyzfile="file.xyz", vmd="vmd")

Applies the periodic boundary conditions on the cellulose crystal over xy-plane using the number of fragments (e.g. chains) as parameter. It returns cellulose XYZ file needed to prepare the PDB and PSF files.

CelluloseBuilder.pbcdata — Method

pbcdata(xsize::Int64, ysize::Int64, zsize::Int64)

Dictionary useful for (a, b, c) definition of the crystal.

CelluloseBuilder.pbcdata — Method

pbcdata(monomers::Int64; chains=nothing)

Dictionary useful for cellulose definition of the layers and fibril.

CelluloseBuilder.translate — Method

translate(phase::String)

CelluloseBuilder.translate — Method

translate(x::Vector{Float64}, y::Vector{Float64}, z::Vector{Float64}, phase::String)

CelluloseBuilder.translate — Method

translate(coords::Vector{Vector{Float64}}, phase::String)

Transforms the coordinates of the asymmetric unit cell to attend the translational symmetry of the cellulose phase (Iα, Iβ, II or III). This function returns the fractional coordinates of the unit cell.

CelluloseBuilder.writePDB — Method

writePDB(monomers::Int64, pdbnames::Vector{String}; phase="Iβ", covalent=true, check_inversion=false, topology_file=DEFAULT_CARB_TOPOLOGY_FILE, vmd="vmd", vmdDebug=true)

Writes the final PSF and PDB files from the cellulose fragments.

CelluloseBuilder.writeXYZ — Method

writeXYZ(atoms::Vector{Vector{String}}, x::Vector{Vector{Float64}}, y::Vector{Vector{Float64}}, z::Vector{Vector{Float64}}; vmd="vmd", natoms=nothing, xyzfile=nothing)
writeXYZ(atoms::Vector{String}, x::Vector{Float64}, y::Vector{Float64}, z::Vector{Float64}; vmd="vmd", natoms=nothing, xyzfile=nothing)
writeXYZ(atoms::Vector{String}, xyzcoords::Vector{Vector{Float64}}; vmd="vmd", natoms=nothing, xyzfile=nothing)

This function is able to save the atomic information (labels + cartesian coordinates) on XYZ file.

CelluloseBuilder.z_expansion — Method

z_expansion(atoms::Vector{String}, x::Vector{Float64}, y::Vector{Float64}, z::Vector{Float64}, zsize::Int64; phase="Iβ")

This function is able to propagate the crystalline system across the z-axis.